8276743: Make openjdk build Zip Archive generation "reproducible"

+++ make/jdk/src/classes/build/tools/makezipreproducible/MakeZipReproducible.java
@@ -21,7 +21,7 @@
  * questions.
  */
 
-package build.tools.generatezip;
+package build.tools.makezipreproducible;
 
 import java.io.*;
 import java.nio.file.*;
@@ -36,24 +36,24 @@ import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
 
 /**
- * Generate a zip file in a "reproducible" manner from the input files or
- * directory.
+ * Generate a zip file in a "reproducible" manner from the input zip file.
  * Standard zip tools rely on OS file list querying whose ordering varies
  * by platform architecture, this class ensures the zip entries are ordered
  * and also supports SOURCE_DATE_EPOCH timestamps.
  */
-public class GenerateZip {
+public class MakeZipReproducible {
+    String input_file = null;
     String fname = null;
     String zname = "";
     long   timestamp = -1L;
-    List<String> files = new ArrayList<>();;
     boolean verbose = false;
 
-    Set<File> entries = new LinkedHashSet<>();
+    // Keep a sorted Set of ZipEntrys to be processed, so that the zip is reproducible
+    SortedMap<String, ZipEntry> entries  = new TreeMap<String, ZipEntry>();
 
     private boolean ok;
 
-    public GenerateZip() {
+    public MakeZipReproducible() {
     }
 
     public synchronized boolean run(String args[]) {
@@ -67,26 +67,28 @@ public class GenerateZip {
                 zname = zname.substring(2);
             }
 
-            if (verbose) System.out.println("Files or directories to zip: "+files);
+            if (verbose) System.out.println("Input zip file: " + input_file);
 
-            File zipFile = new File(fname);
-            // Check archive to create does not exist
-            if (!zipFile.exists()) {
-                // Process Files
-                for(String file : files) {
-                    Path filepath = Paths.get(file);
-                    processFiles(filepath);
-                }
-
-                try (FileOutputStream out = new FileOutputStream(fname)) {
-                    boolean createOk = create(new BufferedOutputStream(out, 4096));
-                    if (ok) {
-                        ok = createOk;
-                    }
-                }
-            } else {
-                error("Target zip file "+fname+" already exists.");
+            File inFile  = new File(input_file);
+            if (!inFile.exists()) {
+                error("Input zip file does not exist");
                 ok = false;
+            } else {
+                File zipFile = new File(fname);
+                // Check archive to create does not exist
+                if (!zipFile.exists()) {
+                    // Process input ZipEntries
+                    ok = processInputEntries(inFile);
+                    if (ok) {
+                        try (FileOutputStream out = new FileOutputStream(fname)) {
+                            ok = create(inFile, new BufferedOutputStream(out, 4096));
+                        }
+                    } else {
+                    }
+                } else {
+                    error("Target zip file "+fname+" already exists.");
+                    ok = false;
+                }
             }
         } catch (IOException e) {
             fatalError(e);
@@ -126,8 +128,13 @@ public class GenerateZip {
                         return false;
                     }
                 } else {
-                    // file or dir to zip
-                    files.add(args[count]);
+                    // input zip file
+                    if (input_file != null) {
+                        error("Input zip file already specified");
+                        usageError();
+                        return false;
+                    }
+                    input_file = args[count];
                 }
                 count++;
             }
@@ -139,13 +146,13 @@ public class GenerateZip {
             return false;
         }
         if (fname == null) {
-            error(String.format("-f <archiveName> must be specified"));
+            error("-f <outputArchiveName> must be specified");
             usageError();
             return false;
         }
         // If no files specified then default to current directory
-        if (files.size() == 0) {
-            error("No input directory or files were specified");
+        if (input_file == null) {
+            error("No input zip file specified");
             usageError();
             return false;
         }
@@ -153,119 +160,64 @@ public class GenerateZip {
         return true;
     }
 
-    // Walk tree matching files and adding to entries list
-    void processFiles(Path path) throws IOException {
-        File fpath = path.toFile();
-        boolean pathIsDir = fpath.isDirectory();
-
-        // Keep a sorted Set of files to be processed, so that the Jmod is reproducible
-        // as Files.walkFileTree order is not defined
-        SortedMap<String, Path> filesToProcess  = new TreeMap<String, Path>();
-
-        Files.walkFileTree(path, Set.of(FileVisitOption.FOLLOW_LINKS),
-            Integer.MAX_VALUE, new SimpleFileVisitor<Path>() {
-                @Override
-                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
-                    throws IOException
-                {
-                    Path relPath;
-                    String name;
-                    if (pathIsDir) {
-                        relPath = path.relativize(file);
-                        name = relPath.toString();
-                    } else {
-                        relPath = file;
-                        name = file.toString();
-                    }
-                    filesToProcess.put(name, file);
-                    return FileVisitResult.CONTINUE;
-                }
-        });
-
-        // Process files in sorted order
-        for (Map.Entry<String, Path> entry : filesToProcess.entrySet()) {
-            String name = entry.getKey();
-            Path   filepath = entry.getValue();
-
-            File f = filepath.toFile();
-            entries.add(f);
+    // Process input zip file and add to sorted entries set
+    boolean processInputEntries(File inFile) throws IOException {
+        try (FileInputStream fis = new FileInputStream(inFile);
+             ZipInputStream  zis = new ZipInputStream(fis)) {
+            ZipEntry entry;
+            while ((entry = zis.getNextEntry()) != null) {
+                entries.put(entry.getName(), entry);
+            }
         }
+
+        return true;
     }
 
     // Create new zip from entries
-    boolean create(OutputStream out) throws IOException
+    boolean create(File inFile, OutputStream out) throws IOException
     {
-        try (ZipOutputStream zos = new ZipOutputStream(out)) {
-            for (File file: entries) {
-                addFile(zos, file);
+        try (ZipFile zipFile = new ZipFile(inFile);
+             ZipOutputStream zos = new ZipOutputStream(out)) {
+            for (Map.Entry<String, ZipEntry> entry : entries.entrySet()) {
+                ZipEntry zipEntry = entry.getValue();
+                if (zipEntry.getSize() > 0) {
+                    try (InputStream eis = zipFile.getInputStream(zipEntry)) {
+                        addEntry(zos, zipEntry, eis);
+                    }
+                } else {
+                    addEntry(zos, zipEntry, null);
+                }
             }
         }
         return true;
     }
 
-    // Ensure a consistent entry name format
-    String entryName(String name) {
-        name = name.replace(File.separatorChar, '/');
-
-        if (name.startsWith("/")) {
-            name = name.substring(1);
-        } else if (name.startsWith("./")) {
-            name = name.substring(2);
-        }
-        return name;
-    }
-
-    // Add File to Zip
-    void addFile(ZipOutputStream zos, File file) throws IOException {
-        String name = file.getPath();
-        boolean isDir = file.isDirectory();
-        if (isDir) {
-            name = name.endsWith(File.separator) ? name : (name + File.separator);
-        }
-        name = entryName(name);
-
-        if (name.equals("") || name.equals(".") || name.equals(zname)) {
-            return;
-        }
-
-        long size = isDir ? 0 : file.length();
-
+    // Add Entry and data to Zip
+    void addEntry(ZipOutputStream zos, ZipEntry entry, InputStream entryInputStream) throws IOException {
         if (verbose) {
-            System.out.println("Adding: "+name);
+            System.out.println("Adding: "+entry.getName());
         }
 
-        ZipEntry e = new ZipEntry(name);
-        // Set to specified timestamp if set otherwise use file lastModified time
+        // Set to specified timestamp if set otherwise leave as original lastModified time
         if (timestamp != -1L) {
-            e.setTime(timestamp);
-        } else {
-            e.setTime(file.lastModified());
+            entry.setTime(timestamp);
         }
-        if (size == 0) {
-            e.setMethod(ZipEntry.STORED);
-            e.setSize(0);
-            e.setCrc(0);
-        }
-        zos.putNextEntry(e);
-        if (!isDir) {
-            byte[] buf = new byte[8192];
-            int len;
-            try (FileInputStream fis = new FileInputStream(file);
-                 FileChannel fic = fis.getChannel()) {
-                fic.transferTo(0, fic.size(), Channels.newChannel(zos));
-            }
+
+        zos.putNextEntry(entry);
+        if (entry.getSize() > 0 && entryInputStream != null) {
+            entryInputStream.transferTo(zos);
         }
         zos.closeEntry();
     }
 
     void usageError() {
         error(
-        "Usage: GenerateZip [-v] -f <zip_file> <files_or_directories>\n" +
+        "Usage: MakeZipReproducible [-v] [-t <SOURCE_DATE_EPOCH>] -f <output_zip_file> <input_zip_file>\n" +
         "Options:\n" +
         "   -v  verbose output\n" +
         "   -f  specify archive file name to create\n" +
         "   -t  specific SOURCE_DATE_EPOCH value to use for timestamps\n" +
-        "If any file is a directory then it is processed recursively.\n");
+        "   input_zip_file re-written as a reproducible zip output_zip_file.\n");
     }
 
     void fatalError(Exception e) {
@@ -277,7 +229,7 @@ public class GenerateZip {
     }
 
     public static void main(String args[]) {
-        GenerateZip z = new GenerateZip();
+        MakeZipReproducible z = new MakeZipReproducible();
         System.exit(z.run(args) ? 0 : 1);
     }
 }
